scsi: ufs: Port UFS quirks for legacy devices
The new devices seem to be provisioned the right way, hence not
requiring the purge red flag disable havoc prevention quirks.
Legacy devices cannot be reprovisioned though, so that's what
we have.. and we just avoid risking devices if anybody tries to
do the porting on 5.4.

So..........

This is a porting of the quirks/hacks required on Sony boards
featuring UFS cards,
************* reqiired to avoid a full UFS card erase *************
*************        at kernel boot, which would      *************
************* E R A S E   T H E   B O O T L O A D E R *************

.... taken from the current implementation found in both the
Open Devices kernel 4.14 as of the current HEAD (10/07/2019)
and the Sony kernel copyleft release for Sony EDO platform,
version 58.0.A.3.39.

BEWARE: DO NOT BOOT SoMC YOSHINO, TAMA, KUMANO, EDO, LENA WITHOUT
THIS COMMIT!!!!
- BOOTING ONE OF THEM WITHOUT THIS COMMIT WILL PERMANENTLY
- (HARD) BRICK YOUR DEVICE WITH ZERO CHANCES TO RECOVER IT.

** JUST PROBING THE UFS CARD WITHOUT READING OR WRITING TO IT
   WITHOUT THESE QUIRKS WILL PRODUCE HAVOC.
        DO     NOT     EVEN     TRY!!!
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^